// backend/src/modules/music/dto/azuracast-response.dto.ts
/* eslint-disable @typescript-eslint/ban-types */
// Generated by https://quicktype.io
//
// To change quicktype's target language, run command:
//
//   "Set quicktype target language"

export interface AzuraCastResponse {
  links: TopLevelLinks;
  page: number;
  per_page: number;
  rows: RowData[];
  total: number;
  total_pages: number;
}

export interface TopLevelLinks {
  first: string;
  last: string;
  next: string;
  previous: string;
}

export interface RowData {
  is_cover_art: boolean;
  is_dir: boolean;
  links: RowLinks;
  media: Media;
  path: string;
  path_short: string;
  playlists: Playlist[];
  size: number;
  text: string;
  timestamp: number;
}

export interface RowLinks {
  download: string;
  rename: string;
}

export interface Media {
  album: Album;
  art: string;
  art_updated_at: number;
  artist: string;
  custom_fields: any[] | object;
  genre: string;
  id: string;
  is_playable: boolean;
  isrc: string;
  length: number;
  length_text: string;
  links: MediaLinks;
  lyrics: string;
  media_id: number;
  text: string;
  title: string;
  unique_id: string;
}

export type Album = {
  Empty: string;
  MarchNoir: string;
};

export interface MediaLinks {
  art: string;
  edit: string;
  play: string;
  waveform: string;
}

export interface Playlist {
  count: number;
  id: number;
  name: Name;
  short_name: ShortName;
}

export type Name = {
  lofilyAppPlaylist: string;
};

export type ShortName = {
  lofilyappplaylist: string;
};